+++ UI/preview/chord_recordings/wav/A_d.wav
