+++ UI/preview/chord_recordings/wav/C#_d.wav
