+++ UI/preview/chord_recordings/wav/C#_u.wav
