+++ UI/preview/chord_recordings/wav/F#_d.wav
