+++ UI/preview/chord_recordings/wav/F#_u.wav
