+++ UI/preview/chord_recordings/wav/F_u.wav
