+++ UI/preview/chord_recordings/wav/Fm_u.wav
