+++ UI/preview/chord_recordings/wav/G#_d.wav
