+++ UI/preview/chord_recordings/wav/G#_u.wav
