+++ UI/preview/chord_recordings/wav/G#m_d.wav
