+++ UI/preview/chord_recordings/wav/G_d.wav
