+++ UI/preview/chord_recordings/wav/G_u.wav
