+++ UI/preview/chord_recordings/wav/Gm_u.wav
